tweak bundle
