+++ assets/panel/prefs.js
@@ -44,3 +44,4 @@ pref("devtools.debugger.features.shortcuts", true);
 pref("devtools.debugger.project-directory-root", "");
 pref("devtools.debugger.features.root", false);
 pref("devtools.debugger.features.column-breakpoints", false);
+pref("devtools.debugger.features.map-scopes", true);
