+++ src/utils/pause/scopes/types.js
@@ -1,0 +1,14 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
+
+// @flow
+
+import type { BindingContents } from "debugger-html";
+
+export type NamedValue = {
+  name: string,
+  generatedName?: string,
+  path: string,
+  contents: BindingContents | NamedValue[]
+};
